Refactor - make Controller control
This is in preparation for implementing issue #10 which will
provide the option to apply changes directly to the database.

Previously the knowledge that change script application was actually
going to a file is spread all over the codebase.

Introduced the concept of an "Applier" which is the only thing
that actually applies change scripts.  Currently the only
implementation is a PrintStreamApplier.

This change also updates the generated sql to make it a bit more
readable.

For clarity, the generated SQL now includes an explicit
BEGIN TRANSACTION, rather than relying on implicit transactions.
(Except for HSQLDB, which doesn't appear to support explicit
transactions.) This is possibly a breaking change for those
people who put transaction logic in their change script.